forexmind/requirements.txt
# ── Core LLM & Agent ──────────────────────────────────────────────────────────
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0

# ── Broker / Market Data ──────────────────────────────────────────────────────
oandapyV20>=0.7.2
alpha_vantage>=2.3.1
yfinance>=0.2.40
finnhub-python>=2.4.20

# ── HTTP / Async ──────────────────────────────────────────────────────────────
aiohttp>=3.11.0
httpx>=0.28.0
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.3.0

# ── Data Processing ───────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
pandas-ta>=0.3.14b0
scipy>=1.14.0

# ── Machine Learning ──────────────────────────────────────────────────────────
lightgbm>=4.5.0
xgboost>=2.1.0
scikit-learn>=1.6.0
joblib>=1.4.0

# ── Reinforcement Learning ────────────────────────────────────────────────────
stable-baselines3>=2.4.0
gymnasium>=1.0.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# ── Web Interface ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
python-multipart>=0.0.12
jinja2>=3.1.4

# ── CLI & Terminal UI ─────────────────────────────────────────────────────────
rich>=13.9.0
click>=8.1.7
prompt_toolkit>=3.0.48

# ── Telegram Bot ──────────────────────────────────────────────────────────────
python-telegram-bot>=21.9.0

# ── Config & Environment ──────────────────────────────────────────────────────
python-dotenv>=1.0.1
pyyaml>=6.0.2

# ── Visualization ─────────────────────────────────────────────────────────────
plotly>=5.24.0

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10.4

# ── Date / Time ───────────────────────────────────────────────────────────────
python-dateutil>=2.9.0
pytz>=2024.2

# ── NLP / Sentiment ───────────────────────────────────────────────────────────
textblob>=0.18.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
